Project/requirements.txt
#name:      PyHouse/requirements.txt
#author:    D. Brian Kimmel
#contact:   D.BrianKimmel@gmail.com
#copyright: (c) 2018-2018 by D. Brian Kimmel
#created:	2018-09-28
#updated:	2018-09-28
#license:   MIT License
#summary:   For pip in a virtualenv
#
# pip3 install -r venv/PyHouse/Project/requirements.txt

twisted >= 19.10.0
service-identity >= 18.1.0                                                                                                                                                                                          
astral >= 1.10.1
netaddr >= 0.7.19
netifaces >= 0.10.9
jsonpickle >= 1.2
python-dateutil >= 2.8.1
aniso8601 >= 8.0.0
pyudev >= 0.21.0
pyserial >= 3.4.0
PyYAML >= 5.3
ruamel.yaml >= 0.16.6
klein >= 19.6.0
pyopenssl >= 19.1.0
wheel >= 0.33.6

onvif_zeep

### END DBK
                                                                                                                                                                                      